Create image editing code
This is the beginning of my editing program for python. Project started due to joining of newspaper comic team and needing a function to quickly get rid of "fuzz" and bad lighting from no scanner. This code simply makes white space and outlines black- much like the contrast tool in photoshop, but less harsh and proper thresholds are used. Very messy code.